# apps/be/static/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/users/signup:
    post:
      operationId: UsersController_signup
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequestDto'
      responses:
        '201':
          description: ''
      summary: 회원가입
      tags: &ref_1
        - 유저
  /api/users/check-id:
    get:
      operationId: UsersController_checkId
      parameters:
        - name: loginId
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckIdResponseDto'
              examples:
                default:
                  value:
                    available: true
      summary: ID 중복 확인
      tags: *ref_1
  /api/users/find-id:
    get:
      operationId: UsersController_findId
      parameters:
        - name: name
          required: true
          in: query
          schema:
            type: string
        - name: phone
          required: true
          in: query
          schema:
            type: string
        - name: email
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindIdResponseDto'
              examples:
                default:
                  value:
                    loginId: user123
      summary: ID 찾기
      tags: *ref_1
  /api/users/reset-pw:
    post:
      operationId: UsersController_sendResetPw
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPwRequestDto'
      responses:
        '200':
          description: ''
      summary: PW 재설정 이메일 발송
      tags: *ref_1
    patch:
      operationId: UsersController_resetPw
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmResetPwRequestDto'
      responses:
        '200':
          description: ''
      summary: PW 재설정
      tags: *ref_1
  /api/users/reset-pw/verify:
    get:
      operationId: UsersController_verifyResetPw
      parameters:
        - name: token
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: PW 재설정 토큰 유효성 검사
      tags: *ref_1
  /api/users/me:
    get:
      operationId: UsersController_getProfile
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDto'
              examples:
                default:
                  value:
                    id: 1
                    loginId: user123
                    name: 홍길동
                    email: user@example.com
                    phone: '01012345678'
                    role: USER
                    createdAt: '2026-01-15T09:00:00.000Z'
      summary: 프로필 조회
      tags: *ref_1
    patch:
      operationId: UsersController_updateProfile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDto'
              examples:
                default:
                  value:
                    id: 1
                    loginId: user123
                    name: 홍길동
                    email: user@example.com
                    phone: '01012345678'
                    role: USER
                    createdAt: '2026-01-15T09:00:00.000Z'
      summary: 프로필 수정
      tags: *ref_1
...
  /api/reservations:
    get:
      operationId: ReservationsController_findAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationListResponseDto'
              examples:
                default:
                  value:
                    reservations:
                      - id: 1
                        status: CONFIRMED
                        reservationDate: '2026-04-01T14:00:00.000Z'
                        courseId: 1
                        courseName: 은행동 빵지순례 코스
                        driverName: 김기사
                        driverPhone: '01098765432'
                        passengerCount: 2
                        createdAt: '2026-03-28T10:00:00.000Z'
      summary: 예약 목록 조회
      tags: &ref_4
        - 예약
    post:
      operationId: ReservationsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    status: CONFIRMED
                    reservationDate: '2026-04-01T14:00:00.000Z'
                    courseId: 1
                    courseName: 은행동 빵지순례 코스
                    driverName: 김기사
                    driverPhone: '01098765432'
                    passengerCount: 2
                    createdAt: '2026-03-28T10:00:00.000Z'
      summary: 예약 생성
      tags: *ref_4
  /api/reservations/{id}:
    get:
      operationId: ReservationsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    status: CONFIRMED
                    reservationDate: '2026-04-01T14:00:00.000Z'
                    courseId: 1
                    courseName: 은행동 빵지순례 코스
                    driverName: 김기사
                    driverPhone: '01098765432'
                    passengerCount: 2
                    createdAt: '2026-03-28T10:00:00.000Z'
      summary: 예약 상세 조회
      tags: *ref_4
    patch:
      operationId: ReservationsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReservationRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    status: PENDING
                    reservationDate: '2026-04-01T14:00:00.000Z'
                    courseId: 1
                    courseName: 은행동 빵지순례 코스
                    driverName: 김기사
                    driverPhone: '01098765432'
                    passengerCount: 2
                    createdAt: '2026-03-28T10:00:00.000Z'
      summary: 예약 수정
      tags: *ref_4
    delete:
      operationId: ReservationsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: 예약 취소
      tags: *ref_4
  /api/payments:
    post:
      operationId: PaymentsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    reservationId: 1
                    amount: 45000
                    method: CARD
                    status: PAID
                    paidAt: '2026-03-28T10:05:00.000Z'
                    createdAt: '2026-03-28T10:00:00.000Z'
      summary: 결제 요청
      tags: &ref_5
        - 결제
  /api/payments/{id}:
    get:
      operationId: PaymentsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    reservationId: 1
                    amount: 45000
                    method: CARD
                    status: PAID
                    paidAt: '2026-03-28T10:05:00.000Z'
                    createdAt: '2026-03-28T10:00:00.000Z'
      summary: 결제 조회
      tags: *ref_5
  /api/payments/{id}/cancel:
    post:
      operationId: PaymentsController_cancel
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    reservationId: 1
                    amount: 45000
                    method: CARD
                    status: CANCELLED
                    paidAt: '2026-03-28T10:05:00.000Z'
                    createdAt: '2026-03-28T10:00:00.000Z'
      summary: 결제 취소
      tags: *ref_5
...
  /api/notifications/token:
    post:
      operationId: NotificationsController_registerToken
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterFcmTokenDto'
      responses:
        '201':
          description: ''
      summary: FCM 토큰 등록
      tags: *ref_7
  /api/notifications/{id}/read:
    patch:
      operationId: NotificationsController_markAsRead
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: 알림 읽음 처리
      tags: *ref_7
info:
  title: 빵빵 (BreadBread) API
  description: 빵빵 택시 서비스 API 명세서
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    BakerySummaryDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: 르뺑 베이커리
        address:
          type: string
          example: 대전시 중구 은행동 123-4
        lat:
          type: number
          format: double
          example: 36.328
        lng:
          type: number
          format: double
          example: 127.4272
        thumbnailUrl:
          type: string
          example: https://cdn.breadbread.kr/bakeries/lepain.jpg
      required:
        - id
        - name
        - address
        - lat
        - lng
        - thumbnailUrl
...
    SignupRequestDto:
      type: object
      properties:
        loginId:
          type: string
          example: user123
        password:
          type: string
          format: password
          example: password123!
        name:
          type: string
          example: 홍길동
        email:
          type: string
          example: user@example.com
          format: email
        phone:
          type: string
          example: '01012345678'
        verificationToken:
          type: string
          description: 휴대폰 인증 완료 토큰
          example: verify_token_abc123
        role:
          type: string
          enum:
            - USER
            - BUSINESS
            - DRIVER
          description: 가입 역할 (일반 사용자/사업자/기사)
          example: USER
      required:
        - loginId
        - password
        - name
        - email
        - phone
        - verificationToken
        - role
    CheckIdResponseDto:
      type: object
      properties:
        available:
          type: boolean
          example: true
      required:
        - available
    FindIdResponseDto:
      type: object
      properties:
        loginId:
          type: string
          example: user123
      required:
        - loginId
    ResetPwRequestDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: user@example.com
      required:
        - email
    ConfirmResetPwRequestDto:
      type: object
      properties:
        token:
          type: string
          example: reset_token_xyz789
        newPassword:
          type: string
          example: newPassword123!
      required:
        - token
        - newPassword
    UserProfileDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        loginId:
          type: string
          example: user123
        name:
          type: string
          example: 홍길동
        email:
          type: string
          example: user@example.com
        phone:
          type: string
          example: '01012345678'
        role:
          type: string
          enum:
            - USER
            - BUSINESS
            - DRIVER
          example: USER
        createdAt:
          type: string
          format: date-time
          example: '2026-01-15T09:00:00.000Z'
      required:
        - id
        - loginId
        - name
        - email
        - phone
        - role
        - createdAt
    UpdateProfileRequestDto:
      type: object
      properties:
        name:
          type: string
          example: 김철수
        email:
          type: string
          format: email
          example: new@example.com
        phone:
          type: string
          example: '01098765432'
...
    CourseDetailDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: 은행동 빵지순례 코스
        thumbnailUrl:
          type: string
          example: https://cdn.breadbread.kr/courses/yeonnam.jpg
        bakeryCount:
          type: number
          example: 4
        estimatedTime:
          type: number
          description: 예상 소요 시간 (분)
          example: 120
        description:
          type: string
          example: 은행동 핫플 빵집 4곳을 도는 인기 코스입니다.
        bakeries:
          type: array
          items:
            $ref: '#/components/schemas/BakerySummaryDto'
      required:
        - id
        - name
        - thumbnailUrl
        - bakeryCount
        - estimatedTime
        - description
        - bakeries
...
    ReservationSummaryDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        courseName:
          type: string
          example: 연남동 빵지순례 코스
        reservationTime:
          type: string
          format: date-time
          example: '2026-04-10T14:00:00.000Z'
        status:
          type: string
          enum: &ref_8
            - PENDING
            - CONFIRMED
            - CANCELLED
            - COMPLETED
          example: CONFIRMED
        createdAt:
          type: string
          format: date-time
          example: '2026-03-28T10:00:00.000Z'
      required:
        - id
        - courseName
        - reservationTime
        - status
        - createdAt
    ReservationListResponseDto:
      type: object
      properties:
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/ReservationSummaryDto'
      required:
        - reservations
    CreateReservationRequestDto:
      type: object
      properties:
        courseId:
          type: number
          format: int64
          example: 1
        reservationTime:
          type: string
          format: date-time
          example: '2026-04-10T14:00:00.000Z'
        headCount:
          type: number
          minimum: 1
          example: 2
      required:
        - courseId
        - reservationTime
        - headCount
    ReservationDetailDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        courseName:
          type: string
          example: 연남동 빵지순례 코스
        reservationTime:
          type: string
          format: date-time
          example: '2026-04-10T14:00:00.000Z'
        status:
          type: string
          enum: *ref_8
          example: CONFIRMED
        createdAt:
          type: string
          format: date-time
          example: '2026-03-28T10:00:00.000Z'
        course:
          $ref: '#/components/schemas/CourseDetailDto'
        headCount:
          type: number
          example: 2
        totalPrice:
          type: number
          example: 45000
      required:
        - id
        - courseName
        - reservationTime
        - status
        - createdAt
        - course
        - headCount
        - totalPrice
    UpdateReservationRequestDto:
      type: object
      properties:
        reservationTime:
          type: string
          format: date-time
          example: '2026-04-11T15:00:00.000Z'
        headCount:
          type: number
          minimum: 1
          example: 3
    CreatePaymentRequestDto:
      type: object
      properties:
        reservationId:
          type: number
          format: int64
          example: 1
        paymentMethod:
          type: string
          enum:
            - CARD
            - TOSS_PAY
            - STRIPE
          example: TOSS_PAY
        amount:
          type: number
          example: 45000
      required:
        - reservationId
        - paymentMethod
    PaymentDetailDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        reservationId:
          type: number
          example: 1
        amount:
          type: number
          example: 45000
        status:
          type: string
          enum:
            - PENDING
            - SUCCESS
            - FAILED
            - CANCELLED
            - REFUNDED
          example: SUCCESS
        paymentMethod:
          type: string
          example: TOSS_PAY
        paidAt:
          type: string
          format: date-time
          example: '2026-03-28T10:05:00.000Z'
      required:
        - id
        - reservationId
        - amount
        - status
        - paymentMethod
        - paidAt
...
    RegisterFcmTokenDto:
      type: object
      properties:
        token:
          type: string
          example: fCM_token_abc123xyz
      required:
        - token